llm/requirements.txt
aiogram==3.17.0
faiss-cpu==1.9.0.post1
flask==3.1.0
langchain-community==0.3.15
openai==1.60.0
pip-chill==1.0.3
pymupdf==1.25.2
pypdf==5.1.0
torchaudio==2.5.1+cpu
torchvision==0.20.1+cpu
transformers==4.48.1
waitress==3.0.2
